tests/wpt/tests/html/canvas/element/path-objects/2d.path.roundrect.badinput.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.roundrect.badinput</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.badinput</h1>
<p class="desc">roundRect() throws or does not throw errors given the strange inputs.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("roundRect() throws or does not throw errors given the strange inputs.");
_addTest(function(canvas, ctx) {

  ctx.roundRect(0, 0, 100, 100, { foo: "bar" });  //=> DOMPointInit
  ctx.roundRect(0, 0, 100, 100, undefined);       //=> "missing" -> 0
  ctx.roundRect(0, 0, 100, 100, [[]]);            //=> « DOMPointInit »
  ctx.roundRect(0, 0, 100, 100, [[25]]);          //=> « DOMPointInit »
  ctx.roundRect(0, 0, 100, 100, [undefined]);     //=> « DOMPointInit »
  assert_throws_js(TypeError, function() { ctx.roundRect(0, 0, 100, 100, 0n); });
  assert_throws_js(TypeError, function() { ctx.roundRect(0, 0, 100, 100, { x: 0n }); });
  assert_throws_js(TypeError, function() { ctx.roundRect(0, 0, 100, 100, [{ x: 0n }]); });

});
</script>
